+++ pytorch_lightning/metrics/functional/auroc.py
@@ -15,8 +15,9 @@ from distutils.version import LooseVersion
 from typing import Optional, Sequence, Tuple
 
 import torch
+from torchmetrics.classification.checks import _input_format_classification
+from torchmetrics.utilities.enums import DataType
 
-from pytorch_lightning.metrics.classification.helpers import _input_format_classification, DataType
 from pytorch_lightning.metrics.functional.auc import auc
 from pytorch_lightning.metrics.functional.roc import roc
 from pytorch_lightning.utilities import LightningEnum